feat : handle Book now feature

+++ lib/views/main/home_page.dart
@@ -79,7 +79,7 @@ class _HomePageState extends State<HomePage>
       return;
     }
     profileController.changeCurrentUserInfo(profileInfo);
-    myPointsController.changeMyPoints(profileInfo.points);
+    myPointsController.changeMyPoints(profileInfo.points.toDouble());
     profileController.changeIsInitialLoading(false);
   }
